use context to prevent prop drilling for table subcomponents
