+++ web/src/modules/Categories/hooks/useCategories.tsx
@@ -9,11 +9,11 @@ import { useState } from "react";
 
 export function useCategories() {
   const [count, setCount] = useState(0);
-  const pagingController = useState<PagingRequest>({
+  const paging = useState<PagingRequest>({
     limit: 25,
     offset: 0,
   });
-  const [{ limit, offset }] = pagingController;
+  const [{ limit, offset }, setPaging] = paging;
 
   const fetchData = async () => {
     const response = await fetch(
@@ -26,11 +26,52 @@ export function useCategories() {
     return data;
   };
 
+  const goToFirst = () => setPaging((current) => ({ ...current, offset: 0 }));
+  const goToNext = () =>
+    setPaging((current) => ({
+      ...current,
+      offset: current.offset + current.limit,
+    }));
+
+  const goToPrev = () =>
+    setPaging((current) => {
+      if (current.offset >= current.limit) {
+        return {
+          ...current,
+          offset: current.offset - current.limit,
+        };
+      }
+
+      return current;
+    });
+
+  const goToLast = () =>
+    setPaging((current) => {
+      return {
+        ...current,
+        offset: Math.floor(count / current.offset) * current.limit,
+      };
+    });
+
+  const changeRecordsPerPage = (perPage: number) =>
+    setPaging({
+      limit: perPage,
+      offset: 0,
+    });
+
+  const methods = {
+    goToFirst,
+    goToLast,
+    goToNext,
+    goToPrev,
+    changeRecordsPerPage,
+  };
+
   return {
     count,
-    pagingController,
+    pagingController: [paging[0], methods] as const,
     query: useQuery({
-      queryKey: ["mock", "categories"],
+      queryKey: ["mock", "categories", `limit=${limit}`, `offset=${offset}`],
       async queryFn() {
         return await fetchData();
       },
